Add login and reset password components with improved structure and UI
- Introduced `LoginContent` and `ResetPasswordContent` components to encapsulate login and password reset functionality.
- Refactored `LoginPage` and `ResetPasswordPage` to utilize suspense for loading states.
- Enhanced user experience with session expiration handling and improved error messaging.
- Updated UI elements for consistency and better accessibility.
